refactor(cvaos): move the tests into a test subpackage
Issue: The eleven test modules sat at the top level of the world, mixed in
with the eight source modules, which made the package hard to scan. Every
other Archipelago world keeps its tests in a test subpackage, and the flat
layout also meant the tests shipped inside the apworld players download.

Solution: The tests now live in worlds/cvaos/test, with their relative
imports rewritten one level up and a package init providing the usual
WorldTestBase subclass.
