+++ worlds/cvaos/test/__init__.py
@@ -1,0 +1,5 @@
+from test.bases import WorldTestBase
+
+
+class CVAOSTestBase(WorldTestBase):
+    game = "Castlevania - Aria of Sorrow"
