+++ worlds/cvaos/test/test_custom_pickups.py
