+++ worlds/cvaos/test/test_icon.py
